MM added javascript in headArea, now available everywhere

+++ app/views/View_invoiceAdministration1.php
@@ -92,7 +92,7 @@ and open the template in the editor.
 
                         echo '<a class="actionbutton" href="InvoiceAdministration/updateInvoiceHouse1/' . $row['eid'] .'/' . 'one' .'">Update</a>';
                         echo '&nbsp;';
-                        echo '<a class="redButton" href="InvoiceAdministration/deleteInvoice/' . $row['eid'] . '/' . "1" . '">Delete</a>';
+                        echo '<a class="redButton" onclick="return confirm_delete()" href="InvoiceAdministration/deleteInvoice/' . $row['eid'] . '/' . "1" . '">Delete</a>';
                         echo '</td>';
                         echo '</tr>';
                     }
